empfly-attendance-manager-main/requirements.txt
amqp==5.1.1
asgiref==3.5.0
astroid==2.11.5
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cmake==3.22.3
dill==0.3.4
Django==4.0.4
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-environ==0.8.1
django-smtp-ssl==1.0
django-timezone-field==5.0
djangorestframework==3.13.1
dlib==19.21.1
face-recognition==1.3.0
flake8==4.0.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.3
# opencv-python==4.5.5.64
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sentry-sdk==1.5.7
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
cffi==1.15.1
pycparser==2.21
cryptography==38.0.1
django-health-check==3.17.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
# image-optimizer==0.0.9
piexif==1.1.3
Pillow==10.1.0
sqlparse==0.4.4
typing-extensions==4.9.0